ecshopDjango/requirements.txt
Django==4.1.4
diff-match-patch
django-filter
django-import-export
django-mysql
django-simpleui
djangorestframework==3.14.0
pip
setuptools
wheel
django-mdeditor
Pillow
django-debug-toolbar==4.3.0